I have some cleanup to do, but at least the tabs aren't lost on force quit
