+++ Client/Frontend/Browser/TabManager.swift
@@ -218,6 +218,12 @@ class TabManager: NSObject {
             tab.applyTheme()
         }
     }
+    
+    func preserveTabs() {
+        print("backgrounding and preserving")
+        
+        store.preserveTabs(tabs, selectedTab: selectedTab)
+    }
 
     func shouldClearPrivateTabs() -> Bool {
         return profile.prefs.boolForKey("settings.closePrivateTabs") ?? false
@@ -616,6 +622,8 @@ extension TabManager {
                 }
             }
         }
+        
+        print("restoring tabs on startup: \(count)")
         guard count == 0, !AppConstants.IsRunningTest, !DebugSettingsBundleOptions.skipSessionRestore, store.hasTabsToRestoreAtStartup else {
             return
         }
